Integrated gameplay with engine code

+++ branches/unstable/Server/comm/commserver.h
@@ -31,7 +31,7 @@ public:
     ServerMessage nextServerMessage();
 
     //send messages or updates
-    void sendServerMsg(ServerMessage msg, const std::vector<int>& clients);
+    void sendServerMsg(ServerMessage msg, const std::vector<size_t>& clients);
     void sendServerMsg(const ServerMessage& msg);
     void sendUpdate(const UpdateObject& update, const std::vector<int>& clientIDs);
     void sendUpdateToAll(const UpdateObject& update);
